Add support for component-based transforms
The UsdGeomXformable API specifies that transforms are composed of many
component operations, which are specified by xformOpOrder. Previously,
the logic assumed all transforms were 4x4 matrices, which is incomplete.

In this change, a new attribute is added to allow the user to direct
serialization such that transforms will be collapsed into a single matrix.
USD provides a helper function to do just this, which is leveraged in pxr.i.

Future work:
This is the second case where data is collapsed in C++; it seems to be a
common pattern that shouldn't be special cased, but rather serialization
should provide a mechanism to allow users to do this as needed.

For now, these special cases seem appropriate, given the ubiquity of
transforms and vertex color.