Add support for UsdPrimRange
Since USD implements UsdPrimRange following the range-based for pattern along
with boost::iterator_adapter, it is difficult to directly wrap the class to
C# without it feeling like C++.

Instead, the class is wrapped directly but also implements IEnumerable returning
an instance of USD.NET.RangeIterator, which allows it to behave more like C#,
without requiring a C++ implementation. This feels like a good trade off between
maintainability and readability.