Binding cleanup + pointer fix
 * Add support for some pair types
 * Fix TfRefPtr<T> pass by reference

Previously, TfRefPtr<T> was attempting to make a copy of the RefPtr
by calling TfCreateRefPtr, however this is unnecessary when copying
an existing ref ptr value (it was also incorrect as it was trying to
construct a refptr refptr <T> when it needed refptr<T>).

Now the pointer is copied via the copy ctor.