Add support for UpAxis and default it "Y"
Previously, the user was required to set this via the raw USD
API for every new scene, which doesn't make sense given that
we know Unity is Y-up.

This change both sets the value by default when creating a scene
and exposes the metadata as a property on USD.NET.Scene.