Fix PATH variable when building Swig
Previously, the USD python build would be used when generating
swig bindings. Now the python USD build is only used for the phase which
requires python. After that point, the PATH is switched to the non-python
USD build.

This is important because some header files generated from the python
build introduce python dependencies which will break the USD.NET build
when compiling against the non-python USD build.