Add support for prefix in BindArrayType
This is required when multiple array types convert to the same Vt
value type, for example Quaternion[] and List<Quaternion> both convert
to VtQuatfArray. The bug occurs when looking up the "ToVtArray" function,
which is overloaded to return VtQuatfArray, so only one method may be
specified per conversion class.

Now a conversion class may specify ListToVtArray for List<Quaternion> and
ArrayToVtArray for Quaternion[] to disambiguate the two functions.

Perhaps in the future, all conversion functions should be prefixed to avoid
this additional complexity, which was difficult to track down and understand.