Better support for List<intrinsic T> serialization
Previously, List<int> (for example) could be serialized, but would be
deserialized as int[] due to aliasing of the FromVtValue binding.

In this change, the ListFromVtValue pattern has been adopted, fixing
the aliasing issue and allowing correct round trip serialization of
all intrinsic types.

The next steps for this change is to remove the "prefix" parameter and
make it a first class binding function, i.e. add the binding function
BindListType<T>().