Add support for nested primvar namespaces
The following example is now allowed:

  class PrimvarSample : USD.NET.SampleBase {
    [USD.NET.UsdNamespace("skel")]
    [USD.NET.VertexData(3)]
    public int[] jointIndices;
  }

Note that when a nested class is wrapped in a namespace, the
members of that class can still have VertexData attributes, however
the nested namespace will apply after the "primvars" namespace.
For example:

  class Nested : USD.NET.SampleBase {
    [USD.NET.UsdNamespace("skel")]
    [USD.NET.VertexData(3)]
    public int[] jointIndices;
  }

  class Outer : USD.NET.SampleBase {
    [USD.NET.UsdNamespace("nested")]
    public Nested nestedClass;
  }

When Outer is serialized, the jointIndices member will have the
namespace primvars:nested:skel:jointIndices.

Fixes #21