Export relative to the root object
Previously, all parent transforms and objects would be included above the root object during export. Further more, a phantom object would be placed above imported objects.

In this change, only objects below the root are exported. This means that parent transforms are disabled and paths are translated to be relative to the export root.