Do not mutate SdfPaths in C#
While this reduced garbage, it also broke a fundamental property of SdfPath (it is immutable) and resulted in incredibly subtle bugs. Particularly, when an SdfPath was added to a hash map, it could then be mutated, which would result in the identity of the hash key changing.

For now, revert to the garbage-generating behavior for sanity. In the future, a full GC pass will be done to reduce generated garbage.