Don't throw exceptions from Diagnostic Delegate
This was sometimes causing the system to enter an infinite loop, which isn't totally surprising given the constraints on the delegate callbacks.

Now, rather than throwing an exception, simply report the error to Unity directly. This is not ideal since the caller will continue without knowing an error occurred, however it's better than hanging.