Add support for configurable import behavior
Until now, import behavior was hard coded in SceneImporter. With this change, clients can change the behavior of how meshes are imported by either re-implementing the specific mesh import functions or by replacing the IImporters entirely.

This is leveraged also in this change to provide two different mesh import modes: one for streaming animation and one for a complete import.