Guard against SdfPath / std::vector crash
It turns out Swig returns an internal reference to the std::vector when accessing items, which means as soon as the vector is disposed, any items still remaining from the vector are setup to segfault.

In this change, I've attempted to mitigate the risk by no longer holding / returning SdfPathVector objects to the user.

See this issue on the Swig GitHub project page for additional details:
https://github.com/swig/swig/issues/945