Materials: First cut at perfect round-tripping
Previously, many settings would be lost in round-trip. Now these parameters are stored generically as "unity" attributes in USD and re-imported if present.

I say "first cut" because it feels like this is the beginning of a more generic system to preserve all sorts of Unity-specific details for round-tripping. Textures and mesh renderers, currently have the same problem.