USD.NET: Add AccessMask
This adds the ability for an external "white list" to control what members of a given Sample are read when calling Scene.Read. The primary use case for this feature is to only read time-varying data, without complicating existing code with special cases (e.g. importers & exporters).

For this purpose, USD.NET.Scene now also has the ability to populate the current AccessMask with all the time-varying members from a call to Read(). IsPopulatingAccessMask flag was added to drive this functionality.