Export: Do not zero the root transform
The original motivation for this does not match up with production use cases, where the root transform is almost always desired.

In addition, this fixes the case where a Z-up asset is imported and subsequently exported; previously, the root transform would be zeroed and the asset would appear incorrectly rotated by 90 degrees on export.