Streaming: Use a proper sync event
The old code was temporary/hacky and not intended to be a long term solution. The main thread now waits for an AutoResetEvent which is signaled by background threads when results are available.

Note that despite the AutoResetEvent being a static variable, each unique instantiation of ReadAllJob<T> gets their own instance of this variable.