UsdSkel: use a valid root path when skeleton is on root bone
When the skeleton is exported at the root bone (which isn't a good idea because it may result in double transform) the exporter would previously generate a full absolute path for this joint, not relative to the root bone (the path was in fact the full root bone path).

In this change, the exporter now names this joint </>, the absolute root path. This kind of makes sense since the joint paths are interpreted to be relative to the root bone. However, it also introduces the need to handle this special case.