USD.NET: Include root object in find results
Previously, Scene.Find<T>(usdRoot) would not include the usdRoot itself, after this change it does.

This had some knock-on effects, in that the root prim xform would now be animated, but that's not compatible with importing an asset and moving it around the scene at the root. So now the root transform is also excluded during streaming playback (it just wasn't in the list of prims to process previously).

This also means that the root prim now has other Usd behaviors, such as AssemblyRoot, which has always felt like it was needed anyway.