Transforms: Respect AccessMask on import
Previously, the importer would only /sometimes/ respect the access mask when importing transforms, which manifested as a bug where the transform would reset to identity when non-varying and authored on one frame on a mesh.

Now the mask is always consulted, regardless of the underlying prim type on which the transform was authored.