UsdSkel Import: Find nested skeleton relationships
Previously, the UsdSkelSkeleton relationship would only be queried on the SkelRoot prim. Now the search is extended to all children of the root as well.

I wanted to use the logic that is provided by the UsdSkelCache rather than implementing it myself, which meant the UsdSkelCache had to be populated sooner, which is the majority of complication in this change.

The UsdSkelCache is now populated as part of the multi-threaded hierarchy discovery phase, rather than populating it late in the SceneImporter.

Fixes #92.