Playback: Fix data race
Previously, streaming playback was relying on reading and writing bool[] elements to be atomic, however bool arrays are implemented as byte arrays, which means reading and writing is not atomic and the elements are aliased.

This fixes the issue by checking the reference array and removing the "written" bool array entirely.

Fixes #93