Fix the ExportUsdz function to get every archive inner paths relative.
Use the native USD function: UsdUtilsCreateNewARKitUsdzPackage. This
function find all the dependencies for us and archive everything
together.

To ensure that all path were relative, the archive had to be created in
the same folder were the source usd file and it's dependencies was. The
archive is then copied to the destination folder provided.