If material as metallic and roughness, combine them
This is required because the roughness map is assumed to be packed into the metallic alpha or albedo alpha channel.

This change also sets the base "glossiness" to 0.5 when a roughness map is present, otherwise it is always completely glossy. It's unclear what this value should be set to, in terms of matching UsdPreviewSurface to the built-in render pipeline...