Always define attributes
Previously, attributes would only be defined when a non-null value was passed in. With this change (which is motivated by the need to define output attributes for shaders), when writing a null value, the attribute will still be defined.

In the future we should add support to respect the "override" mode and not author attribute defines when enabled.