Initialize USD before changing CWD
Previously, if the USD libs were initialized AFTER the current working directory was changed for USDZ export. This is no ta problem unless you happen to export USDZ as the first USD operation in the Unity process, then USD initialization will fail because the package resolution relies on a relative path starting with "packages/...". While this is the recommended way to resolve package paths via the UPM docs, it is not robust against changes to the current directory.

As a workaround, this change initializes USD before changing the working directory, however it is still not protected against third party changes to CWD. Following up with packman devs on a more robust approach.

In the worst case, we can attempt to initialize USD as soon as the editor launches, or at least capture the CWD at that point for use with initialization.