Correctly import & export mesh & skeleton bounding boxes
Import: Previously, both the center and extent of imported objects were converted, which generated incorrect bounding boxes (too large). While this was not immediately visible in Unity, when exporting an imported USD object, the exported bounds would be too large. This can be verified by forcing the importer to compute the BBox instead of reading it from USD, then export the imported object (the exported box that is computed is correct, the exported bbox that was originally imported from USD is incorrect).

Export: Previously the SkelRoot bounds was computed in world space with a correction to convert from worldspace to local space. However, this transform was not correct. Rather than trying to fix the transform after accumulation, the accumulation is now preformed in local space by temporarily setting the SkelRoot parent to null.