UsdSkel: Bind skeleton at mesh, not root
Previously the UsdSkel skeleton was bound once at the root and this binding was inherited down namespace to every mesh below it. This did not work well with AR Quick Look (the binding was not respected) and it occasionally introduced issues with mesh bindings in the joint hierarchy.

With this change, now each mesh is bound to the skeleton individually. To do this, the binding must be set after the UsdSkelRoot is discovered, which introduced a slightly awkward bit of logic in the exporter. It would be nice to clean this up in the future.