UsdIo: Do not recurse on null objects
Previously, this code would exit early on null csValues before getting to this point. The code int he body of this if statement is intended to recurse into valid objects that can perhaps be serialized, however it is not intended to recurse on null objects. 

This fix exits when a null csValue is encountered, as it did previously before the larger change in behavior.