SkeletonImporter: use UsdSkelSkinningQuery to retrieve joints info
BuildSkinnedMesh now use a UsdSkelSkinningQuery object to retrieve the
joints information of a primitive. The advantage of that is that query
looks for inherited attributes. So if a prim doesn't provide joints
indices or weights but has a parent prim that does, those will be
accessible through the query.

SkelBindingSample wasn't looking for that and so it would cause errors
in some cases cause indices and weights information was not found even
when they exists on parent prim.