WIP: SkeletonImporter: Unroll skinning information when vertices was unrolled
This is to ensure bones and skinning will match the vertices even when
they were unrolled in the MeshImporter.

To do: Animation in a timeline is still broken, there is still something
to fix there.

To think: This whole "unroll" process is very tedious, we may need to
re-think the process of creating a (skinned) mesh in Unity from the Usd
data.