Move usdnet to package (#161) (#249)
This moves the USD.NET API to the package to make it easier to iterate on the bindings and simplify the build pipeline. There's a small cost in compile time in most cases and with the biggest regression on IL2CPP player builds (12s to 1m16s, initial build only). We still think that the benefits overweight the regression.

Building UsdCs now installs the generated bindings in the USD.NET directory in the package.
USD.NET can still be built with CMake for the folks wanting to just use the bindings but not the usd package.

Bringing the API inside the package exposed compile warnings and the bindings have been updated to fix those.