Fix animated mesh properties
Sanitizing mesh data requires the original topology (face count and face indices) to be available at all times.
But to improve deserialization performance when reading animation data the static properties are filtered out and not available to the import logic.

A ConversionState interface has been added to hold Sample states and is stored inside the AccessMask initialized globally for all prims at the first frame of animation.