Fix for HumanFemale.walk
Writing this fix outlined where the true regression is- changing from Lists to Arrays in the SanitizedSample changed the copy behaviour, so the array was being overwritten. I'll try addressing just that regression on a different branch to see if that simpler fix resolves it, but pushing this too as I think it makes the behaviour clearer, if less efficient.