First pass at adding relative path to UsdAsset.
When Unity projects are moved or shared, the full path that is serialized for an imported asset causes lots of errors. This change attempts to, where possible, save only the relative path inside the Unity project folder, but still fallback to the full path when it's not in the project folder. We then also catch cases more cleanly when they don't exist at the stated path.