Improvements to analytics:
- Adds timings to import and export events
- Adds analytics for recorder export events
- changes primMap so that instanceRoots are accessible

I moved the SendEvent calls into the SceneImporter/Exporter files so that they capture re-imports and API calls. This does mean it will be harder/ messier to report whether the calls came from button presses or users using the APIs, as well as whether they are importing to GameObjects or Prefabs.

Known issues:
- point instancer not correctly reported
- not sure how to report issues/ failures exporting from recorder
- on failures in some cases we no longer know the file extension
- EditorAnalytics should be stripped out of runtime
- need some code cleanup