Add some partial fixes for transparency.
It's 90% of the way there, but for some reason the generated materials need a small fix up by Unity that is triggered by making any change to the material in the inspector. I can't figure out what this fix up is doing that fixes it or how to trigger it automatically, so shelving for now.