[USDU-476] Improve error messaging for Apple Silicon users. (#406)
* Improve messaging around Silicon support.

* Change the error to an exception so that Initialization fails.

* Fix for standalone compile errors.

* Improve error messaging so that it works in all versions and included eventual Windows on Arm.

* Clear up some common exceptions on Initialization of plugin on unsupported Editors.

There are a lot of places that use InitUSD.Initialize() and then don't check it was successful, so I'm not sure whether to fix all of these or just this most common one. Perhaps in some cases the extra exceptions are useful?
In the back of my mind I think it would be better to hide all of the functionality on unsupported platforms, ie grey out all of the menu items and the components in the inspector window as that would be much cleaner and more obvious. However I think this introduces some risk, and still can't capture everything (such as calling USD APIs directly in scripts).

* Add additional clarification from PR review.