Fix for Transparency import on URP. (#407)
* Initial experiment of URP support

* Added todo to remove duplicated method

* Cleaned up unused code

* Temporary URP export fix.

I've patched up the standard shader exporter to fix a bug where URP exported materials didn't look correct. This isn't a fully featured URP export but should plug a hole for now.

* Fix for smoothness mapping.

* Add note about smoothness constant value.

* Set the correct shader keyword and remove smoothness from the docs limitations.

* Fix occlusion mapping on export.

* Small refactor to remove redundant code.

* Added URP export to changelog.

* Add some partial fixes for transparency.

It's 90% of the way there, but for some reason the generated materials need a small fix up by Unity that is triggered by making any change to the material in the inspector. I can't figure out what this fix up is doing that fixes it or how to trigger it automatically, so shelving for now.

* Add Material Setup step to fix transparency

* Update CHANGELOG.md

* Remove unused variable.

* Add fix to readme